refactoring: move markdowneditor logic to its own class
